requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.6
attrs==20.3.0
Automat==20.2.0
black==21.4b2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
cssselect==1.1.0
Django==3.2.2
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
drf-yasg==1.20.0
flake8==3.9.1
greenlet==1.0.0
h2==3.2.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
inflection==0.5.1
isort==5.8.0
itemadapter==0.2.0
itemloaders==1.0.4
itypes==1.2.0
Jinja2==3.0.0
jmespath==0.10.0
jsonpointer==2.1
jsonschema==3.2.0
lazy-object-proxy==1.6.0
lxml==4.6.3
MarkupSafe==2.0.0
mccabe==0.6.1
mypy-extensions==0.4.3
numpy==1.20.2
openapi-codec==1.3.2
packaging==20.9
pandas==1.2.4
parsel==1.6.0
pathspec==0.8.1
priority==1.3.0
Protego==0.1.16
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pycparser==2.20
PyDispatcher==2.0.5
pyflakes==2.3.1
pylint==2.8.2
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyrsistent==0.17.3
python-dateutil==2.8.1
pytz==2021.1
queuelib==1.6.1
regex==2021.4.4
requests==2.25.1
rfc3987==1.3.8
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
Scrapy==2.5.0
scrapy-jsonschema==0.5.0
service-identity==18.1.0
simplejson==3.17.2
six==1.16.0
SQLAlchemy==1.4.13
sqlparse==0.4.1
strict-rfc3339==0.7
toml==0.10.2
Twisted==21.2.0
twisted-iocpsupport==1.0.1
uritemplate==3.0.1
urllib3==1.26.4
w3lib==1.22.0
webcolors==1.11.1
wrapt==1.12.1
xlrd==2.0.1
zope.interface==5.4.0
